Rename talondoc.analyze -> talondoc.analyzer
