+++ talondoc/sphinx/__init__.py
@@ -6,7 +6,7 @@ from sphinx.application import BuildEnvironment, Sphinx
 from typing_extensions import TypeGuard
 
 from .. import __version__
-from ..analyze.standalone import analyse_package
+from ..analyzer.standalone import analyse_package
 from ..registry import NoActiveFile, NoActivePackage, NoActiveRegistry
 from ..util.logging import getLogger
 from .domains import TalonDomain
